refactor(state): Preserve state on window close
- Elevates all persistent app state (keys, logs, selections) from ContentViewModel to the AppEnvironment singleton.
- Refactors ContentViewModel into a lightweight UI state controller.
- Decomposes ContentView into smaller, focused subviews (SidebarView, LogListView) to solve performance issues caused by over-observation.
- Centralizes derived data logic (e.g., currentLogEntries) in AppEnvironment.

Resolves: LLOG-10
